+++ examples/python/plot_svm_eips.py
@@ -21,9 +21,7 @@ from sklearn import datasets, svm
 from sklearn.model_selection import cross_val_score
 
 from smac.configspace import ConfigurationSpace
-from smac.epm.uncorrelated_mo_rf_with_instances import (
-    UncorrelatedMultiObjectiveRandomForestWithInstances,
-)
+from smac.epm.rf.rf_mo import MultiObjectiveRandomForest
 from smac.facade.smac_ac_facade import SMAC4AC
 
 # EIPS related
@@ -104,7 +102,7 @@ if __name__ == "__main__":
     model_kwargs = {"target_names": ["loss", "time"], "model_kwargs": {"seed": 1}}
     smac = SMAC4AC(
         scenario=scenario,
-        model=UncorrelatedMultiObjectiveRandomForestWithInstances,
+        model=MultiObjectiveRandomForest,
         rng=np.random.RandomState(42),
         model_kwargs=model_kwargs,
         tae_runner=svm_from_cfg,